nofilter urls. abuse setCustomerTokenCallback
the abusing part with the callback. After the avarda iframe shows "thanks" the delivery selections are still visible which is really confusing. There does not seem to be a nice callback to hook into that is fast enough so we are abusing the token callback to hide the delivery stuff to just show the thanks page which then takes a moment to redirect to back to the real order confirmation page